Continuing... there are still bugs, probably interrupt-related.
git-svn-id: file:///home/ryo_on/repos/gxemul-repo/gxemul-legacy/trunk@3565 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/cpus/memory_arm.c
@@ -25,7 +25,7 @@
  *  SUCH DAMAGE.
  *
  *
- *  $Id: memory_arm.c,v 1.16 2005-10-01 00:22:11 debug Exp $
+ *  $Id: memory_arm.c,v 1.17 2005-10-01 17:50:12 debug Exp $
  */
 
 #include <stdio.h>
@@ -211,7 +211,7 @@ exception_return:
 		return 0;
 
 	fatal("{ arm memory fault: vaddr=0x%08x domain=%i dav=%i ap=%i "
-	    "access=%i user=%i", vaddr, domain, dav, ap, access, user);
+	    "access=%i user=%i", (int)vaddr, domain, dav, ap, access, user);
 
 	if (d_in_use)
 		fatal(" d=0x%08x", d);
